split the quiz subcomponents into separate files for readabillity
