+++ frontend/src/components/quiz/Quiz.jsx
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import './quiz.css'
 import { Container, Segment } from 'semantic-ui-react';
 import { FeedbackContainer } from './Feedback.jsx';
+import { Question } from './Question.jsx'
 
 import { sendData } from '../../helpers.jsx' 
 /**TODO: Make nav buttons stay in the same place?
@@ -175,219 +176,5 @@ export class Quiz extends Component {
 				</Container>
 			);
 		}
-	}
-}
-
-/*Representing a single question in a quiz. Manages the answer selection*/
-/*-1 is the default value for the chosen answer, and is equivalent to no answer chosen*/
-class Question extends Component {
-	constructor(props) {
-		super(props);
-		this.state= {
-			chosenAnswer:-1,
-			number:0,
-			data: this.props.data,
-			firstQuestion:false,
-			lastQuestion:false
-			};
-		this.changeToggle = this.changeToggle.bind(this);
-		this.nextQuestion = this.nextQuestion.bind(this);
-		this.prevQuestion = this.prevQuestion.bind(this);
-	}
-
-	componentDidMount() {
-		this.setState({
-			data:this.props.data,
-			firstQuestion:this.props.firstQuestion,
-			lastQuestion:this.props.lastQuestion,
-		});
-
-		if(this.props.chosen !== undefined) {
-			this.setState({
-				chosenAnswer:this.props.chosen
-			});
-		}
-	}
-
-	changeToggle(id) {
-		/*console.log("Callback from" + id);*/
-		if(id === this.state.chosenAnswer) {
-			this.setState({chosenAnswer:-1});
-		}
-		else {
-			this.setState({chosenAnswer:id});
-		}
-	}
-
-	nextQuestion() {
-		//console.log("internaly chosen: " + this.state.chosenAnswer);
-		this.props.onChange(1,this.state.chosenAnswer);
-	}
-
-	prevQuestion() {
-		//console.log("internaly chosen: " + this.state.chosenAnswer);
-		this.props.onChange(-1, this.state.chosenAnswer);
-	}
-
-	componentWillReceiveProps(nextProps) {
-		this.setState({
-			data:nextProps.data,
-			firstQuestion:nextProps.firstQuestion,
-			lastQuestion:nextProps.lastQuestion,
-			chosenAnswer:nextProps.chosen
-		});
-	}
-
-	render() {
-		/*chechs to see if any of the navbuttons needs adjustment for first/last*/
-		/*Nice pale error red:#EF4E45*/
-		var nextText='Next';
-		var backgroundColor = ''
-		var nextCursor = 'default'
-		var prevCursor = 'default'
-		var visibility = 'visible'
-		
-		/*defining styles within render like this is probably not great
-		 * but for the first draft it's ok*/
-
-		var styleNavButtonPrev= {
-			minHeight:'50px',
-			backgroundColor:'#6c6c6c',
-			color:'#ffffff',
-			minWidth:'50%',
-			position:'relative',
-			display:'flex',
-			alignItems:'center',
-			borderRadius:'10px',
-			visibility:'visible',
-			cursor: 'default'
-		}
-		var styleNavButtonNext= {
-			minHeight:'50px',
-			backgroundColor:'#6c6c6c',
-			color:'#ffffff',
-			minWidth:'50%',
-			position:'relative',
-			display:'flex',
-			alignItems:'center',
-			borderRadius:'10px'
-		}
-
-		/*
-		var style3= {
-			textAlign:'center',
-			width:'100%',
-			MozUserSelect:'none',
-			WebkitUserSelect:'none',
-			MsUserSelect:'none',
-			UserSelect:'none'
-		};
-		*/
-
-		/*chechs to see if any of the navbuttons needs adjustment for first/last*/
-		/*Nice pale error red:#EF4E45*/
-		var nextText='Next';
-		/*var backgroundColor = '#6c6c6c'*/
-		/*var cursor = ''*/
-
-		if(this.state.lastQuestion===true) {
-			nextText='Finish';
-			backgroundColor='#5EBC43';
-		}
-
-		if(this.state.firstQuestion===true) {
-			prevCursor='default';
-		}
-		else {
-			visibility='visible';
-			prevCursor='pointer';
-		}
-
-		return (
-			<div className="styleQuestion" >
-				<div className="styleText" >
-					{this.state.data["text"]}
-				</div>
-				{/*chosen:{this.state.chosenAnswer}*/}
-				<div>
-					{
-						this.state.data.choices.map((choice)=> {
-							return(
-								<Answer
-								key={choice.choice_text}
-								opNr={choice.id}
-								text={choice.choice_text}
-								toggleCallback={this.changeToggle}
-								curOn={this.state.chosenAnswer}/>)
-						})
-					}
-
-					<div className="styleNavButtons" >
-						<button className="styleNavButtonPrev" style={{visibility:visibility, cursor:prevCursor}} onClick={this.prevQuestion} disabled={this.state.firstQuestion}>
-							<div className="navText">Previous</div>
-						</button>
-						<button className="styleNavButtonNext" style={{backgroundColor: backgroundColor}} onClick={this.nextQuestion}>
-							<div className="navText">{nextText}</div>
-						</button>
-					</div>
-				</div>
-			</div>
-		)
-	}
-}
-/*A simple custom button for presenting an answer option. What option is toggled is
- * monitored from the Question owning the answer. The parent question monitores
- * answers via a callback passed through the prop*/
-class Answer extends Component {
-	constructor(props) {
-		super(props);
-		this.state= {
-			chosen:false,
-			idNum:this.props.opNr,
-			text:this.props.text
-			};
-		this.handleChange = this.handleChange.bind(this);
-	}
-
-	handleChange() {
-		this.props.toggleCallback(this.state.idNum);
-	}
-
-	componentWillMount() {
-		if(this.props.curOn === this.state.idNum) {
-			this.setState({chosen:true});
-		}
-		else {
-			this.setState({chosen:false});
-		}
-	}
-
-	componentWillReceiveProps(nextProps) {
-		if(nextProps.curOn === this.state.idNum) {
-			this.setState({chosen:true});
-		}
-		else {
-			this.setState({chosen:false});
-		}
-	}
-
-	render() {
-		var background ='';
-		if(this.state.chosen) {
-			background='#4C7CFF';
-		}
-		else {
-			background='#68B1FF';
-		}
-
-		return (
-			<div className="answerWrapper">
-				<button className="answerStyle1" style={{backgroundColor:background}} onClick={this.handleChange}>
-					<div className="answerStyle2">
-						{this.state.text}
-					</div>
-				</button>
-			</div>
-			)
 	}
 }
